requirements.txt
pandas>=2.0.0
numpy>=1.25.0
pyarrow>=12.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
typing-extensions>=4.7.0
pytest>=7.4.0
yfinance>=0.2.40
